Use pure ES modules
