+++ examples/benchmarks/package.json
@@ -17,6 +17,6 @@
   },
   "devDependencies": {
     "typescript": "^4.3.2",
-    "vite": "^2.6.4"
+    "vite": "^4.0.0"
   }
 }
