+++ examples/stats-widget/package.json
@@ -19,6 +19,6 @@
   },
   "devDependencies": {
     "typescript": "^4.3.2",
-    "vite": "^2.6.4"
+    "vite": "^4.0.0"
   }
 }
