+++ index.html
@@ -4,6 +4,5 @@
     <title>probe.gl tests</title>
   </head>
   <body>
-    <script src='test-browser-bundle.js'></script>
   </body>
 </html>
