--- modules/main/env.js
@@ -1,3 +1,0 @@
-// This file enables: import 'probe.gl/bench'.
-// Note: Must be published using package.json "files" field
-module.exports = require('@probe.gl/env');
