--- modules/main/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "probe.gl",
-  "description": "JavaScript console instrumentation for browser and Node",
-  "license": "MIT",
-  "version": "3.5.4",
-  "keywords": [
-    "javascript",
-    "profiling",
-    "instrumentation",
-    "logging"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/uber-web/probe.gl.git"
-  },
-  "types": "dist/index.d.ts",
-  "main": "dist/es5/index.js",
-  "module": "dist/esm/index.js",
-  "files": [
-    "dist",
-    "src",
-    "env.js"
-  ],
-  "sideEffects": false,
-  "dependencies": {
-    "@babel/runtime": "^7.0.0",
-    "@probe.gl/env": "3.5.4",
-    "@probe.gl/log": "3.5.4",
-    "@probe.gl/stats": "3.5.4"
-  },
-  "gitHead": "95541d68783bfd50cbaf9b2a15b3cb41f1ff8284"
-}
