--- modules/main/test/index.js
@@ -1,3 +1,0 @@
-import './env/get-browser.spec';
-import './utils/get-hi-res-timestamp.spec';
-import './lib/log.spec';
