--- modules/main/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "include": ["src/**/*"],
-  "exclude": ["node_modules"],
-  "compilerOptions": {
-    "composite": true,
-    "rootDir": "src",
-    "outDir": "dist"
-  },
-  "references": [
-    {"path": "../env"},
-    {"path": "../log"},
-    {"path": "../stats"}
-  ]
-}
