+++ modules/seer/package.json
@@ -3,6 +3,7 @@
   "version": "3.5.4",
   "description": "A customizable devtool solution",
   "license": "MIT",
+  "type": "module",
   "author": "Balthazar Gronon <bgronon@gmail.com>",
   "repository": {
     "type": "git",
@@ -17,7 +18,15 @@
     "seer"
   ],
   "types": "dist/index.d.ts",
-  "main": "dist",
+  "main": "dist/index.cjs",
+  "module": "dist/index.js",
+  "exports": {
+    ".": {
+      "require": "./dist/index.cjs",
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    }
+  },
   "files": [
     "index.js",
     "dist"
