+++ modules/seer/src/index.ts
@@ -1,0 +1,1 @@
+export {default as default} from './api';
