+++ modules/stats-widget/package.json
@@ -2,6 +2,7 @@
   "name": "@probe.gl/stats-widget",
   "description": "DOM widgets for testing and profiling",
   "license": "MIT",
+  "type": "module",
   "version": "3.5.4",
   "keywords": [
     "javascript",
@@ -14,8 +15,15 @@
     "url": "https://github.com/uber-web/probe.gl.git"
   },
   "types": "dist/index.d.ts",
-  "main": "dist/es5/index.js",
-  "module": "dist/esm/index.js",
+  "main": "dist/index.cjs",
+  "module": "dist/index.js",
+  "exports": {
+    ".": {
+      "require": "./dist/index.cjs",
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    }
+  },
   "files": [
     "dist",
     "src"
@@ -25,10 +33,7 @@
     "@babel/runtime": "^7.0.0"
   },
   "peerDependencies": {
-    "@probe.gl/stats": "3.5.1"
-  },
-  "devDependencies": {
-    "@probe.gl/stats": "3.5.4"
+    "@probe.gl/stats": "^3.5.0"
   },
   "gitHead": "95541d68783bfd50cbaf9b2a15b3cb41f1ff8284"
 }
