+++ modules/test-utils/test/diff-images.spec.js
@@ -1,5 +1,5 @@
 import test from 'tape-promise/tape';
-import {isBrowser} from 'probe.gl';
+import {isBrowser} from '@probe.gl/env';
 
 import {_diffImages as diffImages} from '@probe.gl/test-utils';
 
@@ -9,54 +9,53 @@ test('diffImage', async t => {
     t.end();
     return;
   }
-  const {resolve} = require('path');
-  const dataDir = resolve(__dirname, './data');
+  const dataDir = './modules/test-utils/test/data';
 
   const TEST_CASES = [
     {
       title: 'identical images',
-      source1: resolve(dataDir, 'icon-marker.png'),
-      source2: resolve(dataDir, 'icon-marker-2.png'),
+      source1: `${dataDir}/icon-marker.png`,
+      source2: `${dataDir}/icon-marker-2.png`,
       testMatch: match => match === 1,
       success: true
     },
     {
       title: 'aliased edges',
-      source1: resolve(dataDir, 'icon-marker.png'),
-      source2: resolve(dataDir, 'icon-marker-blur.png'),
+      source1: `${dataDir}/icon-marker.png`,
+      source2: `${dataDir}/icon-marker-blur.png`,
       success: true
     },
     {
       title: 'aliased edges',
-      source1: resolve(dataDir, 'icon-marker.png'),
-      source2: resolve(dataDir, 'icon-marker-blur.png'),
+      source1: `${dataDir}/icon-marker.png`,
+      source2: `${dataDir}/icon-marker-blur.png`,
       success: true
     },
     {
       title: 'aliased edges',
-      source1: resolve(dataDir, 'icon-marker.png'),
-      source2: resolve(dataDir, 'icon-marker-blur.png'),
+      source1: `${dataDir}/icon-marker.png`,
+      source2: `${dataDir}/icon-marker-blur.png`,
       options: {includeAA: true},
       success: false
     },
     {
       title: 'color difference',
-      source1: resolve(dataDir, 'icon-marker.png'),
-      source2: resolve(dataDir, 'icon-marker-color.png'),
+      source1: `${dataDir}/icon-marker.png`,
+      source2: `${dataDir}/icon-marker-color.png`,
       options: {threshold: 0.5},
       success: true
     },
     {
       title: 'color difference',
-      source1: resolve(dataDir, 'icon-marker.png'),
-      source2: resolve(dataDir, 'icon-marker-color.png'),
+      source1: `${dataDir}/icon-marker.png`,
+      source2: `${dataDir}/icon-marker-color.png`,
       options: {threshold: 0.5, includeEmpty: false},
       success: false
     },
     {
       title: 'non-existent file',
-      source1: resolve(dataDir, 'icon-marker.png'),
-      source2: resolve(dataDir, 'non-existent.png'),
+      source1: `${dataDir}/icon-marker.png`,
+      source2: `${dataDir}/non-existent.png`,
       success: false
     }
   ];
